sign_up/requirements.txt
boto3~=1.17.44
bcrypt~=3.1.7
pymongo~=3.11.0
PyJWT~=2.3.0
pytest~=6.2.5
requests~=2.24.0
py~=1.10.0
attrs~=21.2.0
toml~=0.10.2
botocore~=1.20.44
s3transfer~=0.3.6
jmespath~=0.10.0
six~=1.15.0
pluggy~=1.0.0
iniconfig~=1.1.1
certifi~=2020.6.20
chardet~=3.0.4
urllib3~=1.25.11
python-dateutil~=2.8.1
idna~=2.10
pyparsing~=2.4.7
pycparser~=2.20